Initial Commit
This patch sets up a new project, using a `Browser.Document`.

I'm forever forgetting the `elm-live` syntax so I've added a little `bin`
directory for regular tasks that need to be run. At the moment the app doesn't
do anything at all but we'll soon add a basic little model.