Plot price history and portfolio value over time
Early play-testing has revealed that the previous chart was not very suitable
for high-stakes play. If the portfolio or balance went above about $20 dollars
or so, it would become very difficult to react to minute changes in the price,
as the scaling was off.

This patch remedies this bifold. The first is to plot Portfolio valuation and
Balance over time in one chart, and the second is to place a plot of price over
time thereabove. The ticks have been sped up to every 200ms, and the price
volatility has been tweaked to allow for swings of up to 5 dollars per tick.
This should make the game much more exciting.

I've also added a `reset` button, for good measure.

Closes #1, thank you for the dilligent insight @22a and @houli